backend/requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
bcrypt==4.1.1
python-jose[cryptography]==3.5.0
cryptography==41.0.7
email-validator==2.1.0

# Data Validation
pydantic==2.5.0
pydantic[email]==2.5.0

# AI/ML Models
ultralytics==8.0.216
roboflow==1.1.58
pillow==10.1.0
inference-sdk>=0.9.0

# Environment Variables
python-dotenv==1.0.0

# Production Server (optional)
gunicorn==21.2.0

# Development & Debugging
pytest==7.4.3
pytest-asyncio==0.21.1
